+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Changelog
   generic scene object simulated in V-REP.
 - Added retrieving vision sensor depth buffer via interface to vision sensor
   simulated in V-REP.
+- Added retrieving vision sensor resolution via interface to vision sensor
+  simulated in V-REP.
 - Added retrieving vision sensor near clipping plane via interface to vision
   sensor simulated in V-REP.
 - Added setting vision sensor near clipping plane via interface to vision